chore: reorganize gui package
