+++ src/pysigil/gui/events.py
