+++ src/pysigil/gui/gui_state.py
