+++ src/pysigil/gui/hub.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from collections.abc import Callable
 
-from .core import Sigil
+from ..core import Sigil
 
 _instances: dict[str, Sigil] = {}
