+++ src/pysigil/gui/widgets.py
